refactor(bootstrap): Remove usage of PurgeCSS
Removed PurgeCSS from webpack.production.js
Added dependency to elm-bootstrap and refactored app to match
Added method for computing every possible meeting configuration to find
a match if it exists.
Added a checkbox for calling this method in the Recalculate Schedule
button. This is a dangerous operation that I wish didn't have to exist.