refactor(module) Moved data and input into modules
Also changed the use of a Vec<> to a BTreeMap<> and removed TimeRange's
`collide` method in favor of a custom `Ord` implementation.
TODO: Better encapsulation
Needs testing

My computer crashed, and I lost about a month's worth of work. This is
hopefully a way to recover some of that work, but this is the first
attempt at recreating it after wiping my computer.